\name{lqa-internal}
\alias{aic.loss}
\alias{bic.loss}
\alias{dev.loss}
\alias{gcv.loss}
\alias{squared.loss}
\alias{lambda.check}
\title{Internal lqa functions} 
\description{Internal lqa functions}
\usage{
  aic.loss (pred.obj)
  bic.loss (pred.obj)
  dev.loss (pred.obj)
  gcv.loss (pred.obj)
  squared.loss (pred.obj)
  lambda.check (lambda)
}
\author{Jan Ulbricht}
\details{
  These are not to be called by the user.
}
\keyword{internal}
